Move all routing code to separate project, in line with Go impl
