+++ ProtoActor.sln
@@ -40,6 +40,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Plugins", "examples\Plugins
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Proto.Actor.Tests", "tests\Proto.Actor.Tests\Proto.Actor.Tests.csproj", "{6A8DB932-5230-4179-A3F0-291407E0419D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Proto.Routing", "src\Proto.Routing\Proto.Routing.csproj", "{9486DBD2-4103-4A15-A389-C2287D726791}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -158,6 +160,18 @@ Global
 		{6A8DB932-5230-4179-A3F0-291407E0419D}.Release|x64.Build.0 = Release|x64
 		{6A8DB932-5230-4179-A3F0-291407E0419D}.Release|x86.ActiveCfg = Release|x86
 		{6A8DB932-5230-4179-A3F0-291407E0419D}.Release|x86.Build.0 = Release|x86
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|x64.ActiveCfg = Debug|x64
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|x64.Build.0 = Debug|x64
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|x86.ActiveCfg = Debug|x86
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Debug|x86.Build.0 = Debug|x86
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|x64.ActiveCfg = Release|x64
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|x64.Build.0 = Release|x64
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|x86.ActiveCfg = Release|x86
+		{9486DBD2-4103-4A15-A389-C2287D726791}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -177,5 +191,6 @@ Global
 		{AFF0BC22-B7E5-43C2-AB4F-5BBBF688C3F3} = {59DCCC96-DDAF-469F-9E8E-9BC733285082}
 		{E8DC3998-0C17-4F02-87AE-A2F960867420} = {AFF0BC22-B7E5-43C2-AB4F-5BBBF688C3F3}
 		{6A8DB932-5230-4179-A3F0-291407E0419D} = {9AA2BCF0-19AB-4DD9-8D91-7D188E463806}
+		{9486DBD2-4103-4A15-A389-C2287D726791} = {771514F1-12AE-4A26-89CB-2646D3EF7034}
 	EndGlobalSection
 EndGlobal
