+++ src/Proto.Routing/IHashable.cs
@@ -1,0 +1,7 @@
+namespace Proto.Routing
+{
+    public interface IHashable
+    {
+        string HashBy();
+    }
+}
