+++ src/Proto.Routing/Messages/Routees.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+namespace Proto.Routing
+{
+    public class Routees
+    {
+        public Routees(List<PID> pids)
+        {
+            PIDs = pids;
+        }
+
+        public List<PID> PIDs { get; }
+    }
+}
