+++ src/Proto.Routing/Messages/RouterAddRoutee.cs
@@ -1,0 +1,7 @@
+namespace Proto.Routing
+{
+    public class RouterAddRoutee : RouterManagementMessage
+    {
+        public PID PID { get; set; }
+    }
+}
