+++ src/Proto.Routing/Messages/RouterBroadcastMessage.cs
@@ -1,0 +1,7 @@
+namespace Proto.Routing
+{
+    public class RouterBroadcastMessage : RouterManagementMessage
+    {
+        public object Message { get; set; }
+    }
+}
