+++ src/Proto.Routing/Messages/RouterGetRoutees.cs
@@ -1,0 +1,6 @@
+namespace Proto.Routing
+{
+    public class RouterGetRoutees : RouterManagementMessage
+    {
+    }
+}
