+++ src/Proto.Routing/Messages/RouterManagementMessage.cs
@@ -1,0 +1,6 @@
+namespace Proto.Routing
+{
+    public abstract class RouterManagementMessage
+    {
+    }
+}
