+++ src/Proto.Routing/Messages/RouterRemoveRoutee.cs
@@ -1,0 +1,7 @@
+namespace Proto.Routing
+{
+    public class RouterRemoveRoutee : RouterManagementMessage
+    {
+        public PID PID { get; set; }
+    }
+}
