+++ src/Proto.Routing/Routers/BroadcastGroupRouterConfig.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+
+namespace Proto.Routing.Routers
+{
+    internal class BroadcastGroupRouterConfig : GroupRouterConfig
+    {
+        public BroadcastGroupRouterConfig(params PID[] routees)
+        {
+            Routees = new HashSet<PID>(routees);
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new BroadcastRouterState();
+        }
+    }
+}
