+++ src/Proto.Routing/Routers/BroadcastPoolRouterConfig.cs
@@ -1,0 +1,15 @@
+namespace Proto.Routing.Routers
+{
+    internal class BroadcastPoolRouterConfig : PoolRouterConfig
+    {
+        public BroadcastPoolRouterConfig(int poolSize)
+            : base(poolSize)
+        {
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new BroadcastRouterState();
+        }
+    }
+}
