+++ src/Proto.Routing/Routers/ConsistentHashGroupRouterConfig.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+
+namespace Proto.Routing.Routers
+{
+    internal class ConsistentHashGroupRouterConfig : GroupRouterConfig
+    {
+        public ConsistentHashGroupRouterConfig(params PID[] routees)
+        {
+            Routees = new HashSet<PID>(routees);
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new ConsistentHashRouterState();
+        }
+    }
+}
