+++ src/Proto.Routing/Routers/ConsistentHashPoolRouterConfig.cs
@@ -1,0 +1,15 @@
+namespace Proto.Routing.Routers
+{
+    internal class ConsistentHashPoolRouterConfig : PoolRouterConfig
+    {
+        public ConsistentHashPoolRouterConfig(int poolSize)
+            : base(poolSize)
+        {
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new ConsistentHashRouterState();
+        }
+    }
+}
