+++ src/Proto.Routing/Routers/IGroupRouterConfig.cs
@@ -1,0 +1,6 @@
+namespace Proto.Routing.Routers
+{
+    public interface IGroupRouterConfig : IRouterConfig
+    {
+    }
+}
