+++ src/Proto.Routing/Routers/IPoolRouterConfig.cs
@@ -1,0 +1,6 @@
+namespace Proto.Routing.Routers
+{
+    public interface IPoolRouterConfig : IRouterConfig
+    {
+    }
+}
