+++ src/Proto.Routing/Routers/IRouterConfig.cs
@@ -1,0 +1,9 @@
+namespace Proto.Routing.Routers
+{
+    public interface IRouterConfig
+    {
+        void OnStarted(IContext context, Props props, RouterState router);
+
+        RouterState CreateRouterState();
+    }
+}
