+++ src/Proto.Routing/Routers/RandomPoolRouterConfig.cs
@@ -1,0 +1,15 @@
+namespace Proto.Routing.Routers
+{
+    internal class RandomPoolRouterConfig : PoolRouterConfig
+    {
+        public RandomPoolRouterConfig(int poolSize)
+            : base(poolSize)
+        {
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new RandomRouterState();
+        }
+    }
+}
