+++ src/Proto.Routing/Routers/RoundRobinPoolRouterConfig.cs
@@ -1,0 +1,15 @@
+namespace Proto.Routing.Routers
+{
+    internal class RoundRobinPoolRouterConfig : PoolRouterConfig
+    {
+        public RoundRobinPoolRouterConfig(int poolSize)
+            : base(poolSize)
+        {
+        }
+
+        public override RouterState CreateRouterState()
+        {
+            return new RoundRobinRouterState();
+        }
+    }
+}
