+++ src/Proto.Routing/Routers/RouterState.cs
@@ -1,0 +1,17 @@
+﻿// -----------------------------------------------------------------------
+//  <copyright file="Routing.cs" company="Asynkron HB">
+//      Copyright (C) 2015-2016 Asynkron HB All rights reserved
+//  </copyright>
+// -----------------------------------------------------------------------
+
+using System.Collections.Generic;
+
+namespace Proto.Routing.Routers
+{
+    public abstract class RouterState
+    {
+        public abstract HashSet<PID> GetRoutees();
+        public abstract void SetRoutees(HashSet<PID> routees);
+        public abstract void RouteMessage(object message, PID sender);
+    }
+}
